Creation of Factories and refactored singletons away
Factories allow creation of entities of similar types. Each extends the
BaseFactory which stores the world the factory is made for. Each entity
now has a reference to world in order to have global accessible
functions without static globals or Singletons. This will allow for
multiple simulations to be in parallel now