Restructure and More performance based additions
Resture eliminates duplicate files and folders by moving all resources
into src/resources and delete the other sized images we will not be
using.

The editied files reflect work done to optimise the drawing of the ants
but culling processing for only on screen entities and not drawing images
with rotation after certain zoom levels.

Check diffs for more information