Changed compiler to Eclipse JDT Compiler. Tweeks to GUI.
Turns out that on windows, if you have the JDK but without an
appropriate JAVA_HOME Variable, you cannot use
Tools.getSystemCompiler(). It will just return null.

Because of this I have had to include the use of the JDT batch
compiler. It is all in and working now and just needs testing in
windows.

I have also added proper horizontal scrollbars to elements on in the GUI